Module Sihl_type__Queue_workable_job

type t = {

name : string;

work : input:string option -> (unit, string) Result.t Lwt.t;

failed : unit -> (unit, string) Result.t Lwt.t;

max_tries : int;

retry_delay : Sihl_core.Time.duration;

}

A workable job can process a job instance that is persisted. We can not store the job directly because of the polymorphic type ('a Job.t).

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val retry_delay : t -> Sihl_core.Time.duration

val max_tries : t -> int

val failed : t -> unit -> (unit, string) Result.t Lwt.t

val work : t -> input:string option -> (unit, string) Result.t Lwt.t

val name : t -> string

module Fields : sig ... end

val of_job : 'a Sihl_type.Queue_job.t -> t